requirements.txt
Django>=4.0
python-dotenv>=1.0.0
Pillow>=9.0
python-telegram-bot>=20.0
requests>=2.28.0
django-crispy-forms>=2.0
crispy-bootstrap5>=2023.0
